# frontend/musor/api-docs.json
{"openapi":"3.0.1","info":{"title":"Todo application open api","description":"It`s API. U can delete, create and update tasks","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"tasks","description":"method for tasks"},{"name":"Authentification","description":"sign up and sign in"}],"paths":{"/tasks/{id}":{"get":{"tags":["tasks"],"summary":"get task","description":"return task by id ","operationId":"getTaskById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}}}},"put":{"tags":["tasks"],"summary":"update task","description":"update task by id","operationId":"updateTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}}}},"delete":{"tags":["tasks"],"summary":"delete task","description":"delete task by id","operationId":"deleteTask","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/tasks":{"get":{"tags":["tasks"],"summary":"get all tasks (with pagenation)","description":"page (start with 0), size, sort","operationId":"getAllTasks","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageTaskResponse"}}}}}},"post":{"tags":["tasks"],"summary":"create new task","description":"adding new task to db","operationId":"addTask","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TaskResponse"}}}}}}},"/auth/register":{"post":{"tags":["Authentification"],"summary":"sign up","description":"register new user","operationId":"registerNewUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/auth/refresh":{"post":{"tags":["Authentification"],"summary":"refresh token","description":"get new access token by refresh token","operationId":"refreshToken","parameters":[{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/auth/logout":{"post":{"tags":["Authentification"],"summary":"logout","description":"logout and forget refresh token","operationId":"logout","parameters":[{"name":"Authorization","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/auth/login":{"post":{"tags":["Authentification"],"summary":"login","description":"if user is authenticated","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/tasks/{id}/toggle":{"patch":{"tags":["tasks"],"summary":"toggle task","description":"toggle task by id","operationId":"toggleTaskPatch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"TaskRequest":{"required":["topic"],"type":"object","properties":{"topic":{"maxLength":100,"minLength":1,"type":"string"},"shortDescription":{"type":"string"},"longDescription":{"type":"string"},"importance":{"maximum":10,"minimum":1,"type":"integer","format":"int32"},"deadline":{"type":"string","format":"date-time"}}},"TaskResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"topic":{"type":"string"},"shortDescription":{"type":"string"},"importance":{"type":"integer","format":"int32"},"createDate":{"type":"string","format":"date-time"},"endDate":{"type":"string","format":"date-time"},"deadline":{"type":"string","format":"date-time"},"complete":{"type":"boolean"}}},"UserRequest":{"required":["password","username"],"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"tgId":{"type":"string"},"avatar":{"type":"string"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"PageTaskResponse":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/TaskResponse"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/SortObject"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"pageNumber":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"pageSize":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"},"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/SortObject"}}},"SortObject":{"type":"object","properties":{"unsorted":{"type":"boolean"},"sorted":{"type":"boolean"},"empty":{"type":"boolean"}}}}}}